@@ -10,20 +10,16 @@
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
+		"@sveltejs/adapter-auto": "^3.3.1",
+		"@sveltejs/kit": "^2.7.3",
 		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@types/bun": "^1.1.12",
-		"svelte": "^5.0.0",
-		"svelte-check": "^4.0.0",
-		"typescript": "^5.0.0",
-		"vite": "^5.0.3"
+		"svelte": "^5.1.3",
+		"svelte-check": "^4.0.5",
+		"typescript": "^5.6.3",
+		"vite": "^5.4.10"
 	},
 	"dependencies": {
-		"@trpc/client": "^10.45.2",
-		"@trpc/server": "^10.45.2",
-		"delay": "^6.0.0",
-		"marked": "^14.1.3",
-		"trpc-sveltekit": "^3.6.2"
+		"marked": "^14.1.3"
 	}
 }
