@@ -1,7 +1,16 @@
-<script>
-   export let data;
-   const { markdownContent } = data;
-   
+<script lang="ts">
+   interface Props {
+      data: {
+         markdownContent: string;
+      };
+   }
+
+   let { data }: Props = $props();
+   let { markdownContent } = data;
 </script>
 
-{@html markdownContent}
+{#if markdownContent}
+    {@html markdownContent}
+{:else}
+    <p>Content not found</p>
+{/if}
